Restructure entry point  - put architecture dependent code in separate files
