+++ src/entry_point/start_item_mock.rs
@@ -1,0 +1,1 @@
+// Empty file needed to provide a correct start_item module
